+++ test/GCHP/integration/intDiffTest.sh
@@ -1,0 +1,1 @@
+../../shared/intDiffTest.sh