+++ test/GCHP/integration/integrationTestCreate.sh
@@ -141,6 +141,7 @@ printf "\nCopying run scripts to: ${itRoot}/${SCRIPTS_DIR}\n"
 cp -f ${envFile}                            ${envDir}/gchp.env
 cp -f ${thisDir}/integrationTest*.sh        ${scriptsDir}
 cp -f ${thisDir}/commonFunctionsForTests.sh ${scriptsDir}
+cp -f ${thisDir}/intDiffTest.sh             ${scriptsDir}
 cp -f ${thisDir}/README.md                  ${scriptsDir}
 cp -f ${thisDir}/README.testroot.md         ${itRoot}/README.md